RavenDB-7406 - ported tests from 3.x
